+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [Refactor : optimize project and code](https://github.com/Tencent/spring-cloud-tencent/pull/506)
 - [Fix typo & Code optimization](https://github.com/Tencent/spring-cloud-tencent/pull/507)
 - [Bugfix：BeanFactoryUtils returns all beans including beans defined in ancestor bean factories](https://github.com/Tencent/spring-cloud-tencent/pull/515)
+- [fix:fix only config file metadata can be used in routing bug.](https://github.com/Tencent/spring-cloud-tencent/pull/518)